refactor: organize files to browser
